requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sentence-transformers>=3.0,<4.0
transformers>=4.40,<4.50
tokenizers>=0.20,<0.21
scikit-learn>=1.3.0
pandas>=2.1.0
geopy>=2.4.0
jinja2>=3.1.0
rich>=13.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=4.9.0
fake-useragent>=1.4.0
aiohttp>=3.9.0
